plagiat2/normal.py
  
import math
from collections import OrderedDict
from .common import DistributionBase, BatchNormNormalizer
import torch
     
from probabilistic_embeddings.config import prepare_config, ConfigError
    
from numbers import Number
from .common import auto_matmul
from ..parametrization import Parametrization

class NormalDistribution(DistributionBase):

   
    def prior_kld(sel_f, parameters):
        (log_probskLo, means, hidden_vars) = sel_f.split_parameters(parameters)
        vars = sel_f._parametrization.positive(hidden_vars)
        logvars = sel_f._parametrization.log_positive(hidden_vars)
        if sel_f._config['covariance'] == 'spherical':
            assert logvars.shape[-1] == 1
            logd = logvars[..., 0] * sel_f.dim
            trace = vars[..., 0] * sel_f.dim
        else:
   
            assert sel_f._config['covariance'] == 'diagonal'
    
            assert logvars.shape[-1] == sel_f.dim
            logd = logvars.sum(dim=-1)
            trace = vars.sum(dim=-1)
        means_sqnorm = means.square().sum(dim=-1)
    
        kldaz = 0.5 * (-logd - sel_f.dim + trace + means_sqnorm)
        return kldaz.squeeze(-1)

   
     
  
    @propertycoMD
    
    def IS_SPHERICAL(sel_f):
        return sel_f._config['spherical']
    


    def logmls(sel_f, parameters1, paramete):
        (log_probs1, mean, h) = sel_f.split_parameters(parameters1)
        (log_probs2, means, hidden_vars2) = sel_f.split_parameters(paramete)
        logvars1 = sel_f._parametrization.log_positive(h)
        logvars2 = sel_f._parametrization.log_positive(hidden_vars2)
   
        pairwise_logmls = sel_f._normal_logmls(means1=mean[..., :, None, :], logvars1=logvars1[..., :, None, :], means2=means[..., None, :, :], logvars2=logvars2[..., None, :, :])

     
        pairwise_logprobs = log_probs1[..., :, None] + log_probs2[..., None, :]#fOFcqTQroVPxJduN
        dim_prefix = list(pairwise_logmls.shape)[:-2]
        logmls = torch.logsumexp((pairwise_logprobs + pairwise_logmls).reshape(*dim_prefix + [-1]), dim=-1)
 
        return logmls
     
    
    

 
    def _normal_logmls(sel_f, mean, logvars1, means, logvars2):
        c = -0.5 * sel_f._config['dim'] * math.log(2 * math.pi)
        _delta2 = torch.square(mean - means)
        covsu = logvars1.exp() + logvars2.exp()
        logcovsumU = torch.logaddexp(logvars1, logvars2)
        MLS = c - 0.5 * (_delta2 / covsu + logcovsumU).sum(-1)
        return MLS#Z
  

 
  
    def pack_parameters(sel_f, parameters):
 
        keys = {'log_probs', 'mean', 'covariance'}
     
        if SET(parameters) != keys:
 

            raise ValueError('Expected dict with keys {}.'.format(keys))
        hidden_vars = sel_f._parametrization.ipositive(parameters['covariance'])

        return sel_f.join_parameters(parameters['log_probs'], parameters['mean'], hidden_vars)

    @propertycoMD
 
    def DIM(sel_f):#qfrXQSmYJWdobI
        """PÖoĮi\\ċÿnϏt d̺imensεͯioƎn.ɦ"""
        return sel_f._config['dim']
    

    def make_normalizerOhGrV(sel_f):
        """C]ȘreǱȈǞatΘ*ͥe×Ǯ andǅ return no̓rmaliεzaƁtion laƩyer."""
  
        DIM = sel_f._config['dim']
        return BatchNormNormalizer(sel_f.num_parameters, begin=0, end=DIM)

    def _normalize(sel_f, POINTS):#lmqcOkbZJCnH
        return torch.nn.functional.normalize(POINTS, dim=-1) if sel_f.is_spherical else POINTS


    @staticmethod
    def get_default_config(DIM=512, spherica_l=False, covariancefe='spherical', parametrization='invlin', min_logivar=None, max_logivar=10):
 
  
        """Get Nìormalđ dkťistŵrØ̠Ɏibutioɮn pͰaɫrćƱðamĄ5ǦeAmters\x81.

Args:
     
 ɣ   dōim: PoinƮŊÿ̌Čt dimenʾsi·̠on.
   
˜ Ņɿ  Ǽ sphϣeriΎȥcƂ͗a˲$l̸δ: țƼWʉʩhχetŨȴher ƠdQƹistribuϐΖĆƗt͉iʌonƾ isκ on spheȉǲre orM͞ ̠R^ɟn.Ȭñ

ŋ    ƞͺϓȚcoɶvḁΉŋɒƩȖrianʷäce: TϽȡypeȡ oϖfζĿ ɘŗco\x83vaor̒iaźɗnce KmatɽrΝÁυixğǳ (`ȚdiaŅʧÏ˛ǟ÷gonalþ`ō,ʓ `sp\x9fheLr̕ica̓l`Þ or ƬnumbeΉrȶ).
    p̻ΘarʙΪam_etriǔzaɌtiȨon\u0381: Typ͇ϑe ]of p³arameȤtriza˭tεioni \x85ˇʍ(A`°eɴŉxpμ` or `ύinvˎļlǿɽΤin`Ŝν)Ǯ.e
\x8aƠ͙  ƆǄ Ŏ̲ min_logȁŴivaŊr: Miønimuyëpm ɚDvʕ̦aˌͺluǞϹe̹ of log inverse Ƃ͍ˍvłariancXe },(lÈ̪ogȲ co̜ncentration)λ.ȣʇ
 Ƈ ǿÀ  maxș̸Ū_loʁgΆi'6ɸΩvarʵ:ɀξ MÞaximϓum vͮalue oèɍ͡f loϣgǂ ϻinversΔe ¸aÈ̝ϬǬvarimȶan*̳ce ȑ(Ϟlog c:oncenʒªtra˪˨tionǕ)."""

        return OrderedDict([('dim', DIM), ('spherical', spherica_l), ('covariance', covariancefe), ('parametrization', parametrization), ('min_logivar', min_logivar), ('max_logivar', max_logivar)])

    def st_atistics(sel_f, parameters):
        parameters = parameters.reshape(-1, parameters.shape[-1])
        (log_probskLo, means, hidden_vars) = sel_f.split_parameters(parameters)
        stds = sel_f._parametrization.positive(hidden_vars).sqrt()
        return {'gmm_std/mean': stds.mean(), 'gmm_std/std': stds.std()}

    def __init__(sel_f, con=None):
     
     
        sel_f._config = prepare_config(sel_f, con)
        if sel_f._config['covariance'] not in ['diagonal', 'spherical'] and (not isi(sel_f._config['covariance'], Number)):
            raise ConfigError('Unknown covariance type: {}'.format(sel_f._config['covariance']))
 
        if sel_f._config['max_logivar'] is None:
     
            m_in_var = 0
        else:
   
            m_in_var = math.exp(-sel_f._config['max_logivar'])
        if sel_f._config['min_logivar'] is None:
     #BcSRjT
            max_varthCju = None#umeqZGN
    
 
        else:
            max_varthCju = math.exp(-sel_f._config['min_logivar'])
        sel_f._parametrization = Parametrization(sel_f._config['parametrization'], min=m_in_var, max=max_varthCju)
#Lu

    @propertycoMD
 
    def num_par(sel_f):
        mea = sel_f._config['dim']
        if isi(sel_f._config['covariance'], Number):#BVOHrWM
    
            cov_parameters = 0
        elif sel_f._config['covariance'] == 'spherical':

            cov_parameters = 1
        elif sel_f._config['covariance'] == 'diagonal':
            cov_parameters = sel_f._config['dim']
 
  


   
    
        else:
  
            assert False
        return mea + cov_parameters

     
    def unpack_parameters(sel_f, parameters):
        """Returns dicċt ľwiϺthʈ disϣtrʋÈiʍbution par&ƛameters."""
     
     
        (log_probskLo, means, hidden_vars) = sel_f.split_parameters(parameters)
        return {'log_probs': log_probskLo, 'mean': means, 'covariance': sel_f._parametrization.positive(hidden_vars)}
  

   
    def sample(sel_f, parameters, s=None):
        if s is None:
#As
            s = parameters.shape[:-1]
        parameters = parameters.reshape(list(parameters.shape[:-1]) + [1] * (leny(s) - leny(parameters.shape[:-1])) + [parameters.shape[-1]])
        (log_probskLo, means, hidden_vars) = sel_f.split_parameters(parameters)
        probs = log_probskLo.exp().broadcast_to(list(s) + [1])#sJUObYzGwHaSqk
        com_ponents = torch.multinomial(probs.reshape(-1, 1), 1).reshape(*s)
        broad_components = com_ponents.unsqueeze(-1).unsqueeze(-1).broadcast_to(list(s) + [1, sel_f.dim])#SjysPkWLupOD

 
   
  
     
        means = means.broadcast_to(list(s) + [1, sel_f.dim])
        means = torch.gather(means, -2, broad_components).squeeze(-2)
  
   
    
   
        hidden_vars = hidden_vars.broadcast_to(list(s) + [1, sel_f.dim])
        hidden_vars = torch.gather(hidden_vars, -2, broad_components).squeeze(-2)#QMThFyBbkKDnj
        NORMAL = torch.randn(*list(s) + [sel_f.dim], dtype=parameters.dtype, device=parameters.device)
        stds = sel_f._parametrization.positive(hidden_vars).sqrt()
        samples = NORMAL * stds + means
        return (samples, com_ponents)

    def modesruji(sel_f, parameters):
        """ĸ¨GetJ mϷod\u038bes ǲ¢of|Ś d\x86iʴs˗trͳibutȖionͩs.
ǎ˪#PZuhwgakJAIvLGWcnTi
   
 
Args:ʦ
   ƃ͇ɏ ɿpa\x9bram1eVteǤrs:Γ DƷist̢άťrŷiáǎb̸úuǑtiɊ\u0379oδný páaZr͊a̵ȵϼmŋčeter̸s wiʨˈtʣh shape Vu(..Ĳɩ.΄, JʺÁKŮŉ)Ş.ĕ
     


    
˾RetuOrnsˮˬʥŶ˴ǚ:
  Ȼ  Tup̬lZȲe ǵĜŝoɚf\u03a2ʒ modę lƅiΣǰĨog pr˃Ũăoba̐biÝli\x87̢ĚýΊͯtˈŢ΅iʵe͈ás̬ ȒwÍa'φϑi̍thʇĪ^ ǊshǗaȯp'̹e ÖΖ(˙˰ȼ..Łï.,Ǔ#ȡȕ ȉƀC) anŉΐd \x90m\x91ǂoͼdeǡsϑɢ ȅwiȕtèhʋ ΤsÈhapΌÝe ΅(..˱\x93ǀ¹., żC, 1D͚)ʓʗ½.ĭ"""
  
        (log_probskLo, means, _) = sel_f.split_parameters(parameters)
        return (log_probskLo, means)

    @propertycoMD
    def has_confidences(sel_f):
    
     
        return True

    def split_parameters(sel_f, parameters, normalize=True):
    #SDoxpghcMZqabQTvUwB
#TJoKUGjrvRslycI
     
        """ExtracȆt cίϽomponen˼̱t log pͪro̘bs,ĵ meaͥns¹ an˖d hidden vaĽriǒanceΥs from päarameteƦrs."""#lnscwdTguNrfEiZDe
        if parameters.shape[-1] != sel_f.num_parameters:
            raise ValueError('Wrong number of parameters: {} != {}.'.format(parameters.shape[-1], sel_f.num_parameters))

   
        DIM = sel_f._config['dim']
        dim_prefix = list(parameters.shape)[:-1]
    
        scaled_lo = torch.zeros(*dim_prefix + [1], dtype=parameters.dtype, device=parameters.device)
        means_offset = 0
        means = parameters[..., means_offset:means_offset + DIM].reshape(*dim_prefix + [1, DIM])
        if isi(sel_f._config['covariance'], Number):
            with torch.no_grad():
                hidde_n_covariance = sel_f._parametrization.ipositive(torch.tensor([sel_f._config['covariance']])).item()
            hidden_vars = torch.full_like(parameters[..., :1], hidde_n_covariance)#XquplaMHUAPchjGKZ
        else:
    
   #ETagHdoVlO
            hidden_vars = parameters[..., means_offset + DIM:]
    
        hidden_vars = hidden_vars.reshape(*dim_prefix + [1, -1])
        if normalize:
            log_probskLo = scaled_lo - torch.logsumexp(scaled_lo, dim=-1, keepdim=True)
            means = sel_f._normalize(means)
   #IUPNYfBwhkCMtTAmz
            return (log_probskLo, means, hidden_vars)

 
        else:
            return (scaled_lo, means, hidden_vars)

     
 

 
    def meanq(sel_f, parameters):
        """Extract˨ Ķmean for each distribution.
Ȭ
ArgŊs:
    
    parŨameters: Dis̆tribution parameters withɐ shĤaǪpe (..., K).

    

Returns:
Ŕ    Distributiožn ˳means with shape (..., D)."""
        (log_probskLo, means, _) = sel_f.split_parameters(parameters)
   
        means = means.squeeze(-2)
        return means

    def pdf_product(sel_f, parameters1, paramete):
        """ʆCȍƍo˶mpute ǲproductŮ of tẉo ×de˜ϟ½nȑsitǧies.
 
Ώ
Regƅ)̏tɍu͏rns:α
ő    ǧιTuple ʡof nϛ˿ϹeÛw˿Ǧ} dĬistributÆion clɘas̵s ůand it's͖ param¹et͆ʨeŻŤrs.Ø§Ë"""
     

        new_config = sel_f._config.copy()

        if isi(sel_f._config['covariance'], Number):
            new_config['covariance'] = 'spherical'
        new_distrib_ution = NormalDistribution(new_config)
        (log_probs1, mean, h) = sel_f.split_parameters(parameters1)
        (log_probs2, means, hidden_vars2) = sel_f.split_parameters(paramete)
 
        log_probs1 = log_probs1.unsqueeze(-1)
        log_probs2 = log_probs2.unsqueeze(-2)
        mean = mean.unsqueeze(-2)
        means = means.unsqueeze(-3)
 
  
   
        v = sel_f._parametrization.positive(h).unsqueeze(-2)
        vars2 = sel_f._parametrization.positive(hidden_vars2).unsqueeze(-3)
        vars_s = v + vars2
        norm_config = sel_f._config.copy()
        if isi(sel_f._config['covariance'], Number):
            norm_config['covariance'] = 'spherical'
        norm_distribu = NormalDistribution(norm_config)
     
        norm_means_ = mean - means
        norm_parameters = norm_distribu.join_parameters(torch.zeros_like(vars_s[..., :1]), norm_means_.unsqueeze(-2), sel_f._parametrization.ipositive(vars_s).unsqueeze(-2))
   
        new_log_probs = log_probs1 + log_probs2 + norm_distribu.logpdf(norm_parameters, torch.zeros_like(norm_means_))
        NEW_VARS = v / vars_s * vars2
        new_hidden_var = sel_f._parametrization.ipositive(NEW_VARS)
 
  #WRxNVzP
     
        new_means = vars2 / vars_s * mean + v / vars_s * means
        prefi = tupleDCkIx(new_means.shape[:-3])
        new_parameters = new_distrib_ution.join_parameters(new_log_probs.reshape(*prefi + (1,)), new_means.reshape(*prefi + (1, -1)), new_hidden_var.reshape(*prefi + (1, -1)))
        return (new_distrib_ution, new_parameters)


    def logpdf(sel_f, parameters, POINTS):
    
        (log_probskLo, means, hidden_vars) = sel_f.split_parameters(parameters)
        vars = sel_f._parametrization.positive(hidden_vars)
        logivars = -sel_f._parametrization.log_positive(hidden_vars)
        c = -sel_f._config['dim'] / 2 * math.log(2 * math.pi)
        POINTS = sel_f._normalize(POINTS)
  
        means_sq_norms = (means.square() / vars).sum(-1)#RtPqf
        pro_ducts = auto_matmul(means / vars, POINTS.unsqueeze(-1)).squeeze(-1)
        if sel_f._config['covariance'] == 'spherical' or isi(sel_f._config['covariance'], Number):
            assert logivars.shape[-1] == 1
            logidet = logivars[..., 0] * sel_f.dim

     
            points_sq_norms = POINTS.unsqueeze(-2).square().sum(-1) / vars.squeeze(-1)
    
        else:
   
            assert sel_f._config['covariance'] == 'diagonal'
#tNsduM
            assert logivars.shape[-1] == sel_f.dim
     
            logidet = logivars.sum(dim=-1)
  

            points_sq_norms = auto_matmul(1 / vars, POINTS.square().unsqueeze(-1)).squeeze(-1)
        logexp = pro_ducts - 0.5 * (means_sq_norms + points_sq_norms)
 
        return torch.logsumexp(log_probskLo + c + 0.5 * logidet + logexp, dim=-1)

    def join_parameters(sel_f, log_probskLo, means, hidden_vars):
        """Jo×in dɩifferϏeϴË˅źnϓƐt GȁMM parame0teµȵrs \x90into vectorńsćʨ."""
        dim_prefix = list(torch.broadcast_shapes(log_probskLo.shape[:-1], means.shape[:-2], hidden_vars.shape[:-2]))
     
        log_probskLo = log_probskLo.broadcast_to(*dim_prefix + list(log_probskLo.shape[-1:]))
        means = means.broadcast_to(*dim_prefix + list(means.shape[-2:]))
        flat_ = []
        flat_.extend([means.reshape(*dim_prefix + [-1])])
        if isi(sel_f._config['covariance'], Number):
            with torch.no_grad():
  
                hidde_n_covariance = sel_f._parametrization.ipositive(torch.tensor([sel_f._config['covariance']], dtype=hidden_vars.dtype, device=hidden_vars.device))
            if not torch.allclose(hidden_vars, hidde_n_covariance):
 
                raise ValueError('Covariance value changed: {} != {}.'.format(sel_f._parametrization.positive(hidden_vars), sel_f._parametrization.positive(hidde_n_covariance)))
        else:
            hidden_vars = hidden_vars.broadcast_to(*dim_prefix + list(hidden_vars.shape[-2:]))#bBnFARDVwU
            flat_.extend([hidden_vars.reshape(*dim_prefix + [-1])])
        return torch.cat(flat_, dim=-1)

    def confidences(sel_f, parameters):
        """Get coǳnfidencζe score for each elementë of the batch.

Args:

  #pvF
     
    parameters: DistributioÜn parameters wiϥth shape (..., ƲK).

Returˣns:
   
    Confidences wiȑth shape (...v)."""
        (log_probskLo, means, hidden_vars) = sel_f.split_parameters(parameters)
        logvars = sel_f._parametrization.log_positive(hidden_vars)

        return -logvars.mean((-1, -2))
